# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Library Management System API",
    "description": "A **Mini Library Management System** built with FastAPI, PostgreSQL, and OpenAI.\n\n## Authentication\n\nAll protected endpoints require a **Bearer JWT** obtained through the OAuth login flow:\n\n1. Open `GET /api/v1/auth/login/google` (or `/github`) in your **browser**.\n2. Complete the OAuth consent screen.\n3. You are redirected to the frontend with `?token=<jwt>` in the URL.\n4. Copy the token and paste it into the **Authorize** dialog (🔒 button above).\n\nEvery protected endpoint then works directly from this Swagger UI.\n\n## Roles & Permissions\n\n| Role | Capabilities |\n|------|--------------|\n| **Admin** | Full access including user management |\n| **Librarian** | Create / edit / delete books; manage all loans |\n| **Member** | Browse and search books; borrow and return **own** loans |\n\nNew OAuth users start as **Member**. An Admin must promote them via `PATCH /api/v1/admin/users/{id}/role`.\n\n## AI Features\n\nAll three AI endpoints degrade gracefully — they never return an error due to missing AI configuration. Check the `source` field (`\"openai\"` vs `\"fallback\"`) to see which path was taken.\n",
    "contact": {
      "name": "Library Management System",
      "url": "https://github.com/maherabbass/Library-Management-System"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login/{provider}": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Start OAuth login",
        "description": "Redirects the browser to the OAuth provider's authorization page.\n\n**Open this URL directly in a browser** — it initiates the OAuth redirect flow. After the user grants consent the provider calls `/callback/{provider}`, which issues a JWT and redirects to the frontend.\n\nSupported providers: `google`, `github`.",
        "operationId": "login_api_v1_auth_login__provider__get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "302 redirect to the provider's authorization page.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Unsupported or unknown OAuth provider."
          },
          "503": {
            "description": "OAuth provider is not configured on this server (missing client credentials)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/callback/{provider}": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "OAuth callback — issues JWT",
        "description": "Handles the OAuth provider redirect after the user grants consent.\n\nThis endpoint is called **by the OAuth provider**, not directly by the client. It exchanges the authorization code for a user profile, creates the user in the database on first login, then:\n\n- **With frontend configured:** redirects the browser to `{FRONTEND_URL}/auth/callback?token=<jwt>` so the SPA can store the token.\n- **Without frontend (API-only mode):** returns the `TokenResponse` JSON directly.\n\nThe JWT is valid for `ACCESS_TOKEN_EXPIRE_MINUTES` minutes (default: 60).",
        "operationId": "callback_api_v1_auth_callback__provider__get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JWT Bearer token (or 302 redirect to frontend with `?token=...`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to frontend SPA with `?token=<jwt>` query parameter."
          },
          "400": {
            "description": "Unsupported or unknown OAuth provider."
          },
          "503": {
            "description": "OAuth provider is not configured on this server (missing client credentials)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get current user",
        "description": "Returns the profile of the currently authenticated user decoded from the Bearer token.\n\nUse this endpoint to:\n- Verify a token is valid.\n- Retrieve the user's `id`, `email`, `name`, `role`, and `oauth_provider`.\n\n**Requires:** `Authorization: Bearer <token>`",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "The authenticated user's profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List all users",
        "description": "Returns every user account registered in the system, ordered by creation date.\n\nUse this endpoint to:\n- Review all accounts and their current roles.\n- Find user IDs for the `PATCH /admin/users/{id}/role` endpoint.\n\n**Requires:** Admin role.",
        "operationId": "get_users_api_v1_admin_users_get",
        "responses": {
          "200": {
            "description": "Array of all user profiles.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "type": "array",
                  "title": "Response Get Users Api V1 Admin Users Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "403": {
            "description": "Forbidden — Admin role required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/role": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Change a user's role",
        "description": "Promotes or demotes a user by updating their RBAC role.\n\n**Available roles:**\n\n| Role | Description |\n|------|-------------|\n| `MEMBER` | Default role — can browse and borrow books |\n| `LIBRARIAN` | Can create/edit/delete books and manage all loans |\n| `ADMIN` | Full access including this user management endpoint |\n\n> **Warning:** demoting the last Admin account will lock everyone out of admin features.\n\n**Requires:** Admin role.",
        "operationId": "patch_user_role_api_v1_admin_users__user_id__role_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user profile reflecting the new role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "403": {
            "description": "Forbidden — Admin role required."
          },
          "404": {
            "description": "User not found."
          },
          "422": {
            "description": "Validation error — `role` must be one of `ADMIN`, `LIBRARIAN`, `MEMBER`."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleUpdate": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "The new role to assign to the user: `ADMIN` | `LIBRARIAN` | `MEMBER`.",
            "examples": [
              "LIBRARIAN"
            ]
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "RoleUpdate",
        "example": {
          "role": "LIBRARIAN"
        }
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT Bearer token. Include it in subsequent requests as:\n\n`Authorization: Bearer <access_token>`",
            "examples": [
              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            ]
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Token scheme. Always `\"bearer\"`.",
            "default": "bearer",
            "examples": [
              "bearer"
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "example": {
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxYjJjM2Q0ZS01ZjZhLTdiOGMtOWQwZS0xZjJhM2I0YzVkNmUiLCJleHAiOjE3MDU5MjE2MDB9.signature",
          "token_type": "bearer"
        }
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique user identifier (UUID v4)."
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "User's verified email address."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name from the OAuth provider."
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "RBAC role: `ADMIN` | `LIBRARIAN` | `MEMBER`. New users start as `MEMBER`."
          },
          "oauth_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Provider",
            "description": "OAuth provider used to sign in: `google` or `github`."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "UTC timestamp when the account was created."
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "created_at"
        ],
        "title": "UserResponse",
        "example": {
          "created_at": "2024-01-10T08:00:00Z",
          "email": "alice@example.com",
          "id": "1b2c3d4e-5f6a-7b8c-9d0e-1f2a3b4c5d6e",
          "name": "Alice Smith",
          "oauth_provider": "google",
          "role": "MEMBER"
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "LIBRARIAN",
          "MEMBER"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from `GET /api/v1/auth/callback/{provider}`. Obtain it by completing the OAuth login flow and copying the `token` query parameter from the redirect URL."
      }
    }
  },
...
}